requirements.txt
# Dependencies for FFmpeg-AutoToolkit

# GUI Framework
PyQt6==6.5.1

# Console Interface
absl-py==1.4.0

# HTTP Requests for downloading FFmpeg
requests==2.31.0

# For handling paths and system utilities
pathlib==1.0.1
platformdirs==3.10.0

# For running external processes
subprocess32==3.5.4; sys_platform=="linux" and python_version<"3.9"

# Unit testing and testing framework
pytest==7.4.2
unittest-xml-reporting==3.0.7

# Config parser
configparser==5.3.0